backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
marshmallow==3.21.3
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.16.2
gunicorn==22.0.0
celery==5.4.0
redis==5.0.7
pytest==8.2.2
pytest-flask==1.3.0
